backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
Werkzeug==3.1.3
